Mike:
When INQUIRY, REQUEST_SENSE and REPORT_LUNS are sent by scsi-ml the 
request_bufflen will be the correct value. I think the only place where 
this could be broken is through sg io when the app does not set the 
command up correctly. For that case I do not think LLDs should have to 
carry the burden of knowing the spc formats so they can set resid (it 
also seems silly for every LLD to do this), so the attached patch 
removes it from iscsi_tcp.c.